コードの最適化とエラーチェックの追加
- `Player.cpp`で`loadout_`パラメータを削除し、`WeaponEditorUI`の呼び出しを簡素化
- `Collider.cpp`で重複した`Vector3 axis`の定義を削除
- `CollisionManager.cpp`で未使用の定数定義をコメントアウト
- `FadeController.cpp`で不要な変数`u`を削除
- `GamePlayScene.cpp`でカットシーンの処理を有効化
- `GridStageSelector.cpp`で`UpdateWheel`のパラメータを削除
- `BallisticEffect.cpp`でセグメント回収処理を簡略化
- `Weapon.cpp`で`ToJson`関数を削除
- `AnimationModel.h`でメンバ変数を初期化
- `Wireframe.cpp`で未使用の`index3`をコメントアウト
- `MP3Loader.cpp`で`nChannels`をキャスト
- `WavLoader.cpp`でエラーチェックを追加
- `WinApp.cpp`でエラーチェックを追加
- `SRVManager.cpp`で`descriptorHeap`を`descriptorHeap_`に変更し、エラーチェックを追加
- `Ken4lowEngine.vcxproj`で警告レベルを`Level3`に変更

+++ Project/EngineLayer/Audio/MP3Loader.cpp
@@ -68,7 +68,7 @@ void Mp3Loader::PlaySEAsync(const std::string& fileName, float volume, float pit
 
 			WAVEFORMATEX format = {};
 			format.wFormatTag = WAVE_FORMAT_PCM;
-			format.nChannels = mp3.info.channels;
+			format.nChannels = static_cast<WORD>(mp3.info.channels);
 			format.nSamplesPerSec = mp3.info.hz;
 			format.wBitsPerSample = 16;
 			format.nBlockAlign = format.nChannels * format.wBitsPerSample / 8;
@@ -204,7 +204,7 @@ void Mp3Loader::StreamAudio(const std::string& fileName, float volume, float pit
 	// デコーダから取得した情報をもとに音声フォーマットを構築
 	WAVEFORMATEX format = {};
 	format.wFormatTag = WAVE_FORMAT_PCM;
-	format.nChannels = mp3.info.channels;
+	format.nChannels = static_cast<WORD>(mp3.info.channels);
 	format.nSamplesPerSec = mp3.info.hz;
 	format.wBitsPerSample = 16;
 	format.nBlockAlign = format.nChannels * format.wBitsPerSample / 8;
